+++ app/assets/javascripts/util/ui_helper.js
@@ -723,10 +723,14 @@ var UIHelper = (function () {
 
     };
 
-    var createSelectFieldString = function(selectID, selectName, classes, options, selectedItem, label, createEmptyOption) {
+    var createSelectFieldString = function(selectID, selectName, classes, options, selectedItem, label, createEmptyOption, displayOptionTooltips) {
 
         if (createEmptyOption == null || createEmptyOption == undefined){
             createEmptyOption = false;
+        }
+
+        if (displayOptionTooltips == null || displayOptionTooltips == undefined){
+            displayOptionTooltips = false;
         }
 
         var fieldString = '<select id="' + selectID + '"'
@@ -751,6 +755,10 @@ var UIHelper = (function () {
         for (var i = 0; i < options.length; i++) {
 
             fieldString += '<option ';
+
+            if (displayOptionTooltips){
+                fieldString += 'title="' + options[i].tooltip + '" ';
+            }
 
             if (selectedItem != null && selectedItem.toString().toLowerCase() == options[i].value.toString().toLowerCase()) {
                 fieldString += 'selected="selected" ';
@@ -1134,18 +1142,6 @@ var UIHelper = (function () {
         };
     }
 
-    function cloneConcept(id) {
-
-        return function () {
-
-            params = {id: id};
-
-            $.get(gon.routes.taxonomy_clone_concept_path, params, function (results) {
-                console.log("Clone Concept " + uuid);
-            });
-        };
-    }
-
     function changeConceptState(element, concept_id, conceptText, newState) {
 
         return function (){
@@ -1167,7 +1163,7 @@ var UIHelper = (function () {
                 if (results.state != null) {
 
                     splitter.prepend(UIHelper.generatePageMessage("The state of concept " + conceptText + " was successfully updated.", true, "success"));
-                    TaxonomyModule.tree.reloadTreeStatedView(TaxonomyModule.getStatedView(), false);
+                    TaxonomyModule.tree.reloadTree(TaxonomyModule.getViewParams(), false);
 
                     // if the mapping module has been loaded then refresh the mapping tree
                     if (MappingModule.tree){